backend/requirements.txt
fastapi
uvicorn
langgraph
groq
sentence-transformers
chromadb
upstash-redis
pdfplumber
httpx
pydantic
python-dotenv
python-multipart
